BSRD/requirements.txt
absl-py==2.0.0
aiohttp==3.8.5
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
blobfile==2.0.2
cachetools==5.3.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.6
cmake==3.26.4
contourpy
cupy==12.3.0
cycler==0.12.1
docker-pycreds==0.4.0
einops==0.4.1
ExifRead==3.0.0
fastrlock==0.8.2
filelock==3.12.2
fonttools==4.45.1
frozenlist==1.4.0
fsspec==2023.6.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.23.4
google-auth-oauthlib==1.1.0
grpcio==1.59.2
idna==3.4
imageio==2.31.1
Jinja2==3.1.2
kiwisolver==1.4.5
lazy_loader==0.3
lightning-utilities==0.9.0
lit==16.0.6
lpips==0.1.4
lxml==4.9.3
Markdown==3.5
MarkupSafe==2.1.3
matplotlib
mpi4py
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.8.0.76
packaging==23.1
pathtools==0.1.2
Pillow==9.5.0
protobuf==4.23.4
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodomex==3.18.0
pyDeprecate==0.3.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-lightning==1.5.10
pytorch-msssim==1.0.0
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.21.0
scipy
sentry-sdk==1.29.2
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sympy==1.12
# tensorboard==2.15.1
# tensorboard-data-server==0.7.2
tifffile
# torch==2.0.1
# torchaudio==2.0.2
# torchmetrics==1.1.0
# torchvision==0.15.2
tqdm==4.66.0
triton==2.0.0
typing_extensions==4.6.3
urllib3==2.0.3
Werkzeug==3.0.1
yarl==1.9.2
piq
wandb